requirements.txt
Django==1.7.7
djangorestframework==3.1.1
djangorestframework-jwt==1.4.0
PyJWT==1.0.1
